catalog-service/requirements.txt
pymongo
Flask
python-dotenv
flask-cors
PyJWT
requests
grpcio
grpcio-tools
protobuf